- add kimera-data (initial commit) - add applications example/kimera-app (initial commit) - fixed bugs in kimera-core
