+++ kimera-app/README.md
@@ -1,0 +1,3 @@
+# kimera-app
+
+Project that collects applications in python.
