+++ kimera-app/kimera_app/data/etl/extractors/pandas_csv_basic_extractor/basic.csv
@@ -1,0 +1,4 @@
+Nombre; Apellidos; Edad; Sexo
+Tomás; Cabrera Gonzales; 70; M
+Abdiquin; Suarez Suarez; 32; M
+Denia; Sanchez García; 12; F
