+++ kimera-app/kimera_app/data/etl/extractors/pandas_csv_basic_extractor/basic.py
@@ -1,0 +1,12 @@
+import kimera_data.components.etl as kimera_etl
+
+
+def extract_csv_file():
+    extractor = kimera_etl.extractor().engine('pandas').format('csv').option("filepath_or_buffer", 'basic.csv').option('sep', ';')
+    return extractor.read()
+
+
+if __name__ == '__main__':
+    data = extract_csv_file()
+
+    print(data)
