+++ kimera-app/kimera_app/data/etl/extractors/pandas_csv_multiple_extractor/basicII.csv
@@ -1,0 +1,4 @@
+Nombre; Apellidos; Edad; Sexo
+Denia; Cabrera Gonzales; 70; F
+Abdiquin; Suarez Suarez; 32; M
+Tomás; Sanchez García; 12; M
