+++ kimera-app/kimera_app/data/etl/transformers/pandas_csv_basic_transformer/basicI.csv
@@ -1,0 +1,5 @@
+Nombre; Apellidos; Edad; Sexo;Others
+Tomás; Cabrera Gonzales; 70; M;
+Abdiquin; Suarez Suarez; 31; M;
+Ismael; Sanchez García; 12; F;
+Denia; Sanchez García; 12; C
