+++ kimera-core/README.md
@@ -1,9 +1,3 @@
-<div>
-    <table align="left" width="25%">
-        <a target="_blank" href="https://python.org/"><img src="https://www.python.org/static/img/python-logo.png"/></a>
-    </table>
-</div>
+# kimera core
 
-# kimera-core
-[![Package Status](https://img.shields.io/pypi/status/pandas.svg)](https://test.pypi.org/project/kimera-core/)
-[![Code style: black](https://img.shields.io/badge/code%20style-black-000000.svg)](https://github.com/psf/black)
+Project that collects core utility tools in python.
