+++ kimera-data/kimera_data/components/etl/__init__.py
@@ -1,0 +1,10 @@
+from kimera_data.components.etl.extractor import ETLExtractor
+from kimera_data.components.etl.transformer import ETLTransformer
+
+
+def extractor():
+    return ETLExtractor()
+
+
+def transformer():
+    return ETLTransformer()
