+++ kimera-data/setup.py
@@ -8,14 +8,14 @@ with open("README.md", "r") as fh:
     long_description = fh.read()
 
 setuptools.setup(
-    name="kimera_core",  # Replace with your own username
-    version="0.0.10",
+    name="kimera_data",  # Replace with your own username
+    version="0.0.1",
     author="Ismael Antonio Sarmiento Barberia",
     author_email="ismaelantonio.sarmiento@gmail.com",
-    description="Project that collects core utility tools in python.",
+    description="Project that collects data utility tools in python",
     long_description=long_description,
     long_description_content_type="text/markdown",
-    url="https://github.com/ismael-sarmiento/Kimera/tree/master/kimera-core",
+    url="https://github.com/ismael-sarmiento/kimera-data",
     packages=setuptools.find_packages(),
     classifiers=[
         "Programming Language :: Python :: 3",
