+++ kimera-data/tox.ini
@@ -3,7 +3,7 @@
 envlist = coverage
 
 [testenv]
-commands = coverage run --source=kimera_core\components -m unittest discover
+commands = coverage run --source=components -m unittest discover
            coverage report -m
            coverage html
