+++ kimera-doc/developer/python/tools/Packaging.md
@@ -7,8 +7,9 @@
 
 ## Pipeline
 ```bash
-python -m pip install --user --upgrade setuptools wheel
+python -m pip install --upgrade setuptools wheel
 python setup.py sdist bdist_wheel
-python -m pip install --user --upgrade twine
+python -m pip install --upgrade twine
 python -m twine upload --repository testpypi dist/*
+pip install -i https://test.pypi.org/simple/ kimera-<NAME>
 ```
